Add translations for contacts-filters
