+++ src/frontend/web_application/src/scenes/ContactBook/components/ContactFilters/index.jsx
@@ -3,24 +3,24 @@ import { FormGrid, FormRow, FormColumn, SelectFieldGroup } from '../../../../com
 
 import './style.scss';
 
-const ContactFilters = ({ onSortDirChange, onSortViewChange, sortDir, sortView }) => (
+const ContactFilters = ({ onSortDirChange, onSortViewChange, sortDir, sortView, __ }) => (
   <FormGrid className="m-contact-filters">
     <FormRow>
       <FormColumn size="shrink">
         <SelectFieldGroup
           name="format-view"
-          className="m-contact-filters__select"
-          label="Name format view"
+          className="m-contacts-filters__select"
+          label={__('contacts-filters.format-view.label')}
           value={sortView}
-          options={[{ value: 'given_name', label: 'Firstname, Name' }, { value: 'family_name', label: 'Name, Firstname' }]}
+          options={[{ value: 'given_name', label: __('contacts-filters.format-view.firstname') }, { value: 'family_name', label: __('contacts-filters.format-view.name') }]}
           onChange={onSortViewChange}
         />
       </FormColumn>
       <FormColumn size="shrink">
         <SelectFieldGroup
-          className="m-contact-filters__select"
+          className="m-contacts-filters__select"
           name="order-by"
-          label="Order by"
+          label={__('contacts-filters.order-by.label')}
           value={sortDir}
           options={[{ value: 'ASC', label: 'A -> Z' }, { value: 'DESC', label: 'Z -> A' }]}
           onChange={onSortDirChange}
@@ -33,6 +33,7 @@ const ContactFilters = ({ onSortDirChange, onSortViewChange, sortDir, sortView }
 ContactFilters.propTypes = {
   onSortDirChange: PropTypes.func,
   onSortViewChange: PropTypes.func,
+  __: PropTypes.func,
   sortDir: PropTypes.string,
   sortView: PropTypes.string,
 };
