+++ src/frontend/web_application/src/scenes/ContactBook/components/ContactFilters/style.scss
@@ -1,7 +1,7 @@
 @import '../../../../styles/common';
 @import '../../../../styles/vendor/bootstrap_foundation-sites';
 
-.m-contact-filters {
+.m-contacts-filters {
 
   &__select {
     @include flex-grid-row;
